Modified llm model logic. - Made the loading of the transformer models lazy. - Made the loading of the tokenizer lazy. - Changed the dtype to be the default dtype by default (instead of manually setting it to float16) when loading models from pretrained. - Deepseek and Ministral overwrite this default dtype by setting it to float16, as otherwise they would not fit into memory. - Slighlty adjusted prompt for SMS dataset

+++ src/text/outlier_detection/pyod_odm.py
@@ -26,12 +26,12 @@ class PyODM(OutlierDetectionModel, ABC):
         self.od_model = self._get_model()
 
     def _train(self):
-        data = self.x_train.cpu().numpy()
+        data = self.x_train.float().cpu().numpy()
         self.od_model.fit(data, None)
 
     def _predict(self):
         test = self.x_test
-        decision_function = self.od_model.decision_function(test.cpu().numpy())
+        decision_function = self.od_model.decision_function(test.float().cpu().numpy())
         self.decision_function = decision_function
 
     def _get_predictions(self) -> List[int]:
